fix: ensure all jobs that we consider are completed
From our observations in #107, the github API occasionally returns
job statuses as "in_progress". Since this action only works for
"completed" jobs, this change adds a backoff-and-retry loop in case
any job is encountered with a non-completed status.

This means that the action can now fail (and can now run for a little
while longer - 10 x 2 seconds), in case the github API response
doesn't converge on completed values.